fix(cpp): small changes
Former-commit-id: d4cd390a4c825dc557fa710cebc730d59f83bd48

+++ cpp/src/libloader/PluginListImplements.cpp
@@ -304,7 +304,7 @@ namespace LibLoader {
     inline void loadsAll(const std::vector<std::string> &paths, const std::vector<PluginAuthority> &authorities) noexcept {
         std::lock_guard lk(PluginListManager::getLock());
 
-        for (int i = 0; i < paths.size(); ++i) {
+        for (size_t i = 0; i < paths.size(); ++i) {
             auto &path = paths[i];
             auto authority = authorities[i];
